refactor: add SearchIcon
