--- src/data.ts
@@ -1,129 +1,0 @@
-import { ITracks } from './models'
-
-export const dataTracks: ITracks = {
-  "count": 29,
-  "next": "http://51.250.72.80:8090/catalog/track/all/?page=2",
-  "previous": null,
-  "results": [
-    {
-      "id": 36,
-      "name": "Action Sport Breakbeat",
-      "author": "Winniethemoog",
-      "release_date": "1991-09-07",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 234,
-      "album": "Workout",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Winniethemoog_-_Action_Sport_Breakbeat.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 35,
-      "name": "Hard Metal Intro",
-      "author": "Winniethemoog",
-      "release_date": "1991-09-06",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 255,
-      "album": "Hard Metal",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Winniethemoog__-_Hard_Metal_Intro.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 34,
-      "name": "Adrenelynne",
-      "author": "Tim Kulig",
-      "release_date": "2007-05-14",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 285,
-      "album": "Adrenelynne",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Tim_Kulig_-_Adrenelynne.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 33,
-      "name": "Classical Metal Workout",
-      "author": "-",
-      "release_date": "1991-09-06",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 246,
-      "album": "Workout",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Musiclfiles_-_Classical_Metal_Workout.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 32,
-      "name": "Essence2",
-      "author": "MED",
-      "release_date": "1920-05-03",
-      "genre": "Электронная музыка",
-      "duration_in_seconds": 205,
-      "album": "Essence2",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/MED_-_Essence2.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 31,
-      "name": "5 cents back",
-      "author": "MED",
-      "release_date": "2005-03-03",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 345,
-      "album": "5 cents back",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/MED_-_5_Cents_Back.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 30,
-      "name": "The World Waltz",
-      "author": "Kevin Macleodburn",
-      "release_date": null,
-      "genre": "Рок музыка",
-      "duration_in_seconds": 305,
-      "album": "The Waltzer",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Kevin_Macleodburn_-_The_World_Waltz.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 29,
-      "name": "Feel Good Rock",
-      "author": "Audionautix",
-      "release_date": "2020-02-12",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 312,
-      "album": "FAST",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Audionautix_-_Feel_Good_Rock.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 28,
-      "name": "Carol Of The Bells",
-      "author": "Alexander Nakarada",
-      "release_date": "2022-04-16",
-      "genre": "Рок музыка",
-      "duration_in_seconds": 205,
-      "album": "Carol Of The Bells",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Alexander_Nakarada_-_Carol_Of_The_Bells.mp3",
-      "stared_user": []
-    },
-    {
-      "id": 27,
-      "name": "Hangtime",
-      "author": "Wova",
-      "release_date": "1991-09-06",
-      "genre": "Электронная музыка",
-      "duration_in_seconds": 306,
-      "album": "Hangtime",
-      "logo": null,
-      "track_file": "http://51.250.72.80:8090/django_media/music_files/Wova_-_Hangtime.mp3",
-      "stared_user": []
-    }
-  ]
-}
