+++ src/features/Search/Search.tsx
@@ -1,6 +1,5 @@
 import React, { FC } from 'react'
-
-import iconSearch from './assets/search.svg'
+import SearchIcon from '../../icons/SearchIcon'
 
 import { cnSearch } from './Search.classname'
 import './Search.css'
@@ -11,17 +10,13 @@ type SearchProps = {
 }
 
 export const Search: FC<SearchProps> = ({ value='', onChange }) => {
-  
-  return (
-    <div className={cnSearch()}>
-      <img className="search__svg" src={iconSearch} alt="search"/>
-      <input
-        className="search__text"
-        type="search"
-        placeholder="Поиск"
-        name="search"
-        value={value}
-        onChange={onChange}/>
-    </div>
-  )
+  return <div className={cnSearch()}>
+    <SearchIcon/>
+    <input className={cnSearch('text')}
+      type="search"
+      placeholder="Поиск"
+      name="search"
+      value={value}
+      onChange={onChange}/>
+  </div>
 }
