+++ src/icons/LoopIcon.tsx
