+++ src/icons/NextIcon.tsx
