+++ src/icons/PauseIcon.tsx
