+++ src/icons/TrackIcon.tsx
